moved home test files into their directory
